+++ packages/crashlytics/src/tracing/tracing-provider.ts
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 
-
 import { resourceFromAttributes } from '@opentelemetry/resources';
 import {
   CompositePropagator,