+++ packages/performance-compat/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @firebase/performance-compat
 
+## 0.2.25
+
+### Patch Changes
+
+- [`8e384c9`](https://github.com/firebase/firebase-js-sdk/commit/8e384c947de73c7f74346b1c01640f3515a4ef0e) [#9883](https://github.com/firebase/firebase-js-sdk/pull/9883) - Updated dependencies.
+
+- Updated dependencies [[`8e384c9`](https://github.com/firebase/firebase-js-sdk/commit/8e384c947de73c7f74346b1c01640f3515a4ef0e)]:
+  - @firebase/component@0.7.3
+  - @firebase/logger@0.5.1
+  - @firebase/performance@0.7.12
+  - @firebase/performance-types@0.2.4
+  - @firebase/util@1.15.1
+
 ## 0.2.24
 
 ### Patch Changes